Rename `source_graph.rs` to `inputs.rs`
