+++ crates/oak_db/src/tests.rs
@@ -1,6 +1,6 @@
 mod file;
 mod file_workspace_root;
+mod inputs;
 mod resolver;
 mod root;
-mod source_graph;
 mod test_db;
